backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-dotenv==1.0.0
email-validator==2.1.0

# Database
motor==3.3.2
pymongo==4.6.1

# Machine Learning
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
joblib==1.3.2

# Monitoring & Logging
watchdog==3.0.0

# Network & Firewall
pyroute2==0.7.12

# Utilities
aiofiles==23.2.1
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Web Push Notifications
pywebpush==1.14.0
py-vapid==1.11.0

